criado as limitações de usuarios

+++ project/candidatos/views.py
@@ -1,3 +1,4 @@
+from django.contrib import messages
 from django.shortcuts import redirect, render, get_object_or_404
 from django.contrib.auth.decorators import login_required
 from candidatos.models import Candidato
@@ -7,8 +8,13 @@ from candidatos.forms import CandidatoForm
 def sem_permissao(request):
     return render(request, 'candidatos/sem_permissao.html')
 
+
+
 @login_required
 def index(request):
+    if not request.user.has_perm('candidatos.view_candidato'):
+        return redirect('sem_permissao')
+
     loja = request.GET.get('loja')
     busca = request.GET.get('busca')
    
@@ -27,18 +33,26 @@ def index(request):
         "candidatos": candidatos,
         "lojas": lojas,
     })
+
+@login_required
 def novo_candidato(request):
+    if not request.user.has_perm('candidatos.add_candidato'):
+        return redirect('sem_permissao')
     if request.method == 'POST':
         form =CandidatoForm(request.POST)
         if form.is_valid():
             form.save()
+            messages.success(request,'Candidato cadastrado com sucesso')
             return redirect('index')
         
     else:
         form = CandidatoForm()
     return render(request, 'candidatos/form.html', {"form":form, "titulo": "Novo Candidato"})
 
+@login_required
 def editar_candidado(request,pk):
+    if not request.user.has_perm('candidatos.change_candidato'):
+        return redirect('sem_permissao')
     candidato = get_object_or_404(Candidato, pk=pk)
     if request.method == "POST":
         form = CandidatoForm(request.POST, instance=candidato)
@@ -49,8 +63,10 @@ def editar_candidado(request,pk):
         form = CandidatoForm(instance=candidato)
 
     return render(request, 'candidatos/form.html', {"form":form, "titulo": "Editar Candidato"})
-
+@login_required
 def deletar_candidatos(request, id):
+    if not request.user.has_perm('candidatos.delete_candidato'):
+        return redirect('sem_permissao')
     candidato = get_object_or_404(Candidato, id = id)
     if request.method == "POST":
         candidato.delete()
